Investigate rails conventions around :js tag
Unknown to the author (and not documented anywhere I can find) there is an interesting distinction
between system and feature specs.
System specs are expected to run using a browser (although strangely the rspec:system generator sets driven_by :rack_test)
whereas feature specs are what the author is more used to, where rack_test is the default and is over-ridden on a case-by-case
basis using the js: true annotation to the spec (or group).
This moves the welcome_spec into the feature tests directory to generate discussion about what type the tests should be on this project.